REF: avoid special-casing in SelectN
